DataSource (Front)

+++ front/pages/[user]/a/[sId]/datasets/index.jsx
@@ -1,4 +1,4 @@
-import AppLayout from "../../../../../components/app/AppLayout";
+import AppLayout from "../../../../../components/AppLayout";
 import MainTab from "../../../../../components/app/MainTab";
 import { ActionButton } from "../../../../../components/Button";
 import { unstable_getServerSession } from "next-auth/next";
@@ -21,17 +21,19 @@ export default function DatasetsView({
   const { data: session } = useSession();
 
   const handleDelete = async (datasetName) => {
-    const res = await fetch(
-      `/api/apps/${session.user.username}/${app.sId}/datasets/${datasetName}`,
-      {
-        method: "DELETE",
-        headers: {
-          "Content-Type": "application/json",
-        },
-      }
-    );
-    const data = await res.json();
-    Router.push(`/${session.user.username}/a/${app.sId}/datasets`);
+    if (confirm("Are you sure you want to delete this dataset entirely?")) {
+      const res = await fetch(
+        `/api/apps/${session.user.username}/${app.sId}/datasets/${datasetName}`,
+        {
+          method: "DELETE",
+          headers: {
+            "Content-Type": "application/json",
+          },
+        }
+      );
+      const data = await res.json();
+      Router.push(`/${session.user.username}/a/${app.sId}/datasets`);
+    }
   };
 
   return (
@@ -43,79 +45,83 @@ export default function DatasetsView({
         <div className="flex flex-initial mt-2">
           <MainTab
             app={{ sId: app.sId, name: app.name }}
-            current_tab="Datasets"
+            currentTab="Datasets"
             user={user}
             readOnly={readOnly}
           />
         </div>
-        <div className="flex flex-1">
-          <div className="flex flex-auto flex-col mx-2 sm:mx-4 lg:mx-8 my-4">
-            <Link href={`/${user}/a/${app.sId}/datasets/new`}>
-              <ActionButton disabled={readOnly}>
-                <PlusIcon className="-ml-1 mr-1 h-5 w-5 mt-0.5" />
-                New Dataset
-              </ActionButton>
-            </Link>
+        <div className="w-full max-w-5xl mt-4 mx-auto">
+          <div className="flex flex-1">
+            <div className="flex flex-auto flex-col mx-2 sm:mx-4 lg:mx-8 my-4">
+              <Link href={`/${user}/a/${app.sId}/datasets/new`}>
+                <ActionButton disabled={readOnly}>
+                  <PlusIcon className="-ml-1 mr-1 h-5 w-5 mt-0.5" />
+                  New Dataset
+                </ActionButton>
+              </Link>
 
-            <div className="mt-4">
-              <ul role="list" className="space-y-4 flex-1">
-                {datasets.map((d) => {
-                  return (
-                    <Link
-                      key={d.name}
-                      href={`/${user}/a/${app.sId}/datasets/${d.name}`}
-                      className="block"
-                    >
-                      <div
+              <div className="mt-4">
+                <ul role="list" className="space-y-4 flex-1">
+                  {datasets.map((d) => {
+                    return (
+                      <Link
                         key={d.name}
-                        className="group px-4 py-4 rounded-sm border border-gray-300"
+                        href={`/${user}/a/${app.sId}/datasets/${d.name}`}
+                        className="block"
                       >
-                        <div className="flex items-center justify-between">
-                          <p className="truncate text-base font-bold text-violet-600">
-                            {d.name}
-                          </p>
-                          {readOnly ? null : (
-                            <div className="ml-2 flex flex-shrink-0">
-                              <TrashIcon
-                                className="h-4 w-4 hidden group-hover:block text-gray-400 hover:text-red-700"
-                                onClick={(e) => {
-                                  e.preventDefault();
-                                  handleDelete(d.name);
-                                }}
-                              />
-                            </div>
-                          )}
-                        </div>
-                        <div className="mt-2 sm:flex sm:justify-between">
-                          <div className="sm:flex">
-                            <p
-                              className={classNames(
-                                d.description
-                                  ? "text-gray-700"
-                                  : "text-gray-300",
-                                "flex items-center text-sm text-gray-700"
-                              )}
-                            >
-                              {d.description ? d.description : "No description"}
+                        <div
+                          key={d.name}
+                          className="group px-4 py-4 rounded border border-gray-300"
+                        >
+                          <div className="flex items-center justify-between">
+                            <p className="truncate text-base font-bold text-violet-600">
+                              {d.name}
                             </p>
+                            {readOnly ? null : (
+                              <div className="ml-2 flex flex-shrink-0">
+                                <TrashIcon
+                                  className="h-4 w-4 hidden group-hover:block text-gray-400 hover:text-red-700"
+                                  onClick={(e) => {
+                                    e.preventDefault();
+                                    handleDelete(d.name);
+                                  }}
+                                />
+                              </div>
+                            )}
+                          </div>
+                          <div className="mt-2 sm:flex sm:justify-between">
+                            <div className="sm:flex">
+                              <p
+                                className={classNames(
+                                  d.description
+                                    ? "text-gray-700"
+                                    : "text-gray-300",
+                                  "flex items-center text-sm text-gray-700"
+                                )}
+                              >
+                                {d.description
+                                  ? d.description
+                                  : "No description"}
+                              </p>
+                            </div>
                           </div>
                         </div>
-                      </div>
-                    </Link>
-                  );
-                })}
-              </ul>
-              <div className="px-2 max-w-4xl mt-2">
-                <div className="text-sm text-gray-400 py-2">
-                  Datasets are used as input data to apps (
-                  <span className="rounded-md px-1 py-0.5 bg-gray-200 font-bold">
-                    input
-                  </span>{" "}
-                  block) or few-shot examples to prompt models (
-                  <span className="rounded-md px-1 py-0.5 bg-gray-200 font-bold">
-                    data
-                  </span>{" "}
-                  block).
+                      </Link>
+                    );
+                  })}
+                </ul>
+                <div className="px-2 max-w-4xl mt-2">
+                  <div className="text-sm text-gray-400 py-2">
+                    Datasets are used as input data to apps (
+                    <span className="rounded-md px-1 py-0.5 bg-gray-200 font-bold">
+                      input
+                    </span>{" "}
+                    block) or few-shot examples to prompt models (
+                    <span className="rounded-md px-1 py-0.5 bg-gray-200 font-bold">
+                      data
+                    </span>{" "}
+                    block).
+                  </div>
                 </div>
               </div>
             </div>
